GraphNG: unified legend for BarChart, GraphNG & other uPlot based visualizations (#31175)
* Legend experiment

* Nits
