+++ packages/grafana-ui/src/components/VizLayout/VizLayout.story.tsx
@@ -24,7 +24,7 @@ export const BottomLegend = () => {
   const items = Array.from({ length: legendItems }, (_, i) => i + 1);
 
   const legend = (
-    <VizLayout.Legend position="bottom" maxHeight="30%">
+    <VizLayout.Legend placement="bottom" maxHeight="30%">
       {items.map((_, index) => (
         <div style={{ height: '30px', width: '100%', background: 'blue', marginBottom: '2px' }} key={index}>
           Legend item {index}
@@ -47,7 +47,7 @@ export const RightLegend = () => {
   const items = Array.from({ length: legendItems }, (_, i) => i + 1);
 
   const legend = (
-    <VizLayout.Legend position="right" maxWidth="50%">
+    <VizLayout.Legend placement="right" maxWidth="50%">
       {items.map((_, index) => (
         <div style={{ height: '30px', width: `${legendWidth}px`, background: 'blue', marginBottom: '2px' }} key={index}>
           Legend item {index}
